requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
httpx==1.7.2
pydantic==1.10.10
python-dotenv==1.0.1

# Modelos externos
openai==1.33.0
anthropic==0.5.0
langchain==0.1.0

# Opcional: modelos locais pequenos
transformers==4.34.0
torch==2.2.0
